quarto-site/blog/posts/pdf-lab-reports/requirements.txt
timm
torch
numpy
pillow
matplotlib
transformers
